fix(finance): 持股成本改浮點（碎股的美金成本有小數）
Firstrade 的 VEA 成本 48.42 美元存成整數 48，報酬率就從 16.7% 變 17.8% ——
碎股部位小，取整的相對誤差反而大。同 shares／last_price 都是 Float。
欄位是當天才建的、生產零筆資料，ALTER TYPE 無損。

+++ core/schemas.py
@@ -1452,7 +1452,7 @@ class HoldingPayload(BaseModel):
     quote_symbol: Optional[str] = ""
     last_price: Optional[float] = None
     manual_value: Optional[int] = None
-    cost_total: Optional[int] = None      # 投資成本（該列的幣別）
+    cost_total: Optional[float] = None    # 投資成本（該列的幣別；碎股有小數）
     sort_order: int = 0
     active: bool = True
     note: Optional[str] = None
